Add Tor and SOCKS5 protocols
Tor: This signifies that the rest of the protocol is accomplished routed
through Tor, e.g. "/dns/www.google.com/http/tor" would be connecting to
www.google.com via HTTP through Tor. Note that the details of getting to
Tor are considered to be the domain of the transport being used, e.g. if
the Tor software is installed locally, it would be routed through the
SOCKS5 proxy.

SOCKS5: This signifies that the protocol is accomplished by routing
through a SOCKS5 proxy, via the IP and port which are part of the
address.